Moved PomFileReadActor to processing package, changed behavior on failure
